Only add auth headers to ntfy.sh when set (#1)
My current deployment is failing, with errors in the logs of

```
2025/05/21 17:59:28 ERROR Failed to send notification err="{\"code\":40101,\"http\":401,\"error\":\"unauthorized\",\"link\":\"https://ntfy.sh/docs/publish/#authentication\"}"
```

Looking at ntfy's issues, it appears that this can happen when the
client tries to set auth on an unprotected topic (my topic currently is
unprotected - there's nothing confidential in the messages)

I took a quick hack at adjusting the code to support this and added a
test.

Please be kind - this is my first Golang thing, and **I have not built
the entire app and tried running it with this config** - I've only added
(and run) the extra test.